Update the Raspberry Pi firmware to the current version of today.
Add a file (distinfo) to keep the commit hash and the commit date of each
file to make the future update of these files easier.

This update brings support to the new B+ model and some important bug fixes,
like a fix to the MMC controller clock which fixes some of the issues with
SD card compatibility.